refractor each component to its own file
